requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-modeltranslation==0.19.9
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.1
